Todo Entity unit test
